% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agree.cac3rd_doc.r
\docType{data}
\name{agree.cac3rd}
\alias{agree.cac3rd}
\title{Dataset showing the distribution of 6 raters by psychiatric condition}
\format{
The first column is of character type whereas the remaining columns 
contains integers.
 \describe{
    \item{Gender}{Contains subject's gender (M or F)}
    \item{Depression}{The interval upper bound}
    \item{Pers.Disorder}{Contains the number of raters who labeled the subject as having personal disorder}
    \item{Schizophrenia}{Contains the number of raters who labeled the subject as having schizophrenia}
    \item{Neurosis}{Contains the number of raters who labeled the subject as having Neurosis}
    \item{Other}{Contains the number of raters who labeled the subject as having Other psychiatric condition}
 }
}
\source{
K. Gwet, PhD.
}
\usage{
agree.cac3rd
}
\description{
This dataset contains rating data about 15 subjects or patients whose 
psychiatric condition was evaluated by 6 raters. Each column represents one 
condition and contains the number of raters who diagnosed the patient with
that condition.
}
\keyword{datasets}
