% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agree.contingency_doc.r
\docType{data}
\name{agree.contingency}
\alias{agree.contingency}
\title{Dataset representing a 10x10 contingency table}
\format{
The first column is alphabetic whereas the remaining columns contains
integers.
 \describe{
    \item{Categories}{Contains category's name C1,C2, ..., C10)}
    \item{C1}{Colum C1}
    \item{C2}{Colum C2}
    \item{C3}{Colum C3}
    \item{C4}{Colum C4}
    \item{C5}{Colum C5}
    \item{C6}{Colum C6}
    \item{C7}{Colum C7}
    \item{C8}{Colum C8}
    \item{C9}{Colum C9}
    \item{C10}{Colum C10}
 }
}
\source{
K. Gwet, PhD.
}
\usage{
agree.contingency
}
\description{
This dataset contains a 10x10 contingency table.
}
\keyword{datasets}
