% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreeCAC_doc.r
\docType{data}
\name{agreeCAC}
\alias{agreeCAC}
\title{Ratings of 15 subjects from 4 raters}
\format{
All 5 columns are of alphabetic type.
 \describe{
    \item{Gender}{Contains the subject's gender (M or F)}
    \item{RaterA}{Contains one of ratings a,b,c produced by RaterA}
    \item{RaterB}{Contains one of ratings a,b,c produced by RaterB}
    \item{RaterC}{Contains one of ratings a,b,c produced by RaterC}
    \item{RaterD}{Contains one of ratings a,b,c produced by RaterD}
 }
}
\source{
K. Gwet, PhD.
}
\usage{
agreeCAC
}
\description{
This dataset contains raw ratings of 15 subjects produced by 4 raters named
as RaterA, RaterB, RaterC and RaterD.
}
\keyword{datasets}
