% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agree.coeff2.r
\name{bangdiwala.table}
\alias{bangdiwala.table}
\title{Bangdiwala B coefficient for 2 raters}
\usage{
bangdiwala.table(ratings, conflev = 0.95, N = Inf)
}
\arguments{
\item{ratings}{A square table of ratings (assume no missing ratings).}

\item{conflev}{An optional parameter that specifies the confidence level used for constructing confidence intervals. By default the function assumes the standard value of 95\%.}

\item{N}{An optional parameter representing the finite population size if any. It is used to perform the finite population correction to the standard error. It's default value is infinity.}

\item{weights}{An optional matrix that contains the weights used in the weighted analysis. By default, this parameter contaings the identity weight matrix, which leads to the unweighted analysis.}
}
\value{
A data frame containing the following 5 variables: coeff.name coeff.val coeff.se coeff.ci coeff.pval.
}
\description{
Bangdiwala B coefficient for 2 raters
}
\examples{
#The dataset "cont3x3abstractors" comes with this package. Analyze it as follows:
bangdiwala.fn(cont3x3abstractors) #Yields Scott's Pi coefficient along with precision measures
Bcoeff <- bangdiwala.fn(cont3x3abstractors)$coeff.val #Yields Scott's coefficient alone.
Bcoeff
q <- nrow(cont3x3abstractors) #Number of categories
}
