% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agree.coeff2.r
\name{bangdiwala2RR.fn}
\alias{bangdiwala2RR.fn}
\title{Bangdiwala B coefficient for 2 raters when input dataset is made up of 2 columns of raw data.}
\usage{
bangdiwala2RR.fn(fra.ratings.raw, conflev = 0.95, N = Inf)
}
\arguments{
\item{fra.ratings.raw}{A dataframe with 2 columns of raw ratings.}

\item{conflev}{An optional parameter that specifies the confidence level used for constructing confidence intervals. By default the function assumes the standard value of 95\%.}

\item{N}{An optional parameter representing the finite population size if any. It is used to perform the finite population correction to the standard error. It's default value is infinity.}
}
\value{
A data frame containing the following 9 variables: coeff.name, b1, b2, 
coeff.val, coeff.se, conf.int, p.value, n and name of the weight used.
}
\description{
Bangdiwala B coefficient for 2 raters when input dataset is made up of 2 columns of raw data.
}
\examples{
The dataset cac.ben.gerry comes with this package. Analyze it as follows:
bangdiwala2RR.fn(cac.ben.gerry[,c(3,4)]), using only the last 2 columns.
The result will be following:
coeff.name        pa        pe coeff.val  coeff.se  conf.int     p.val tot.obs
1 Bangdiwala''s B 0.1322314 0.2066116      0.64 0.2158518 (0.159,1) 7.083e-03      11
w.name
1 Identity
}
