% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cac.ben.gerry_doc.R
\docType{data}
\name{cac.ben.gerry}
\alias{cac.ben.gerry}
\title{Ratings of 12 units from 2 raters named Ben and Gerry}
\format{Each row of this dataset describes an interval and the interpretation of the magnitude it represents.
 \describe{
    \item{Group}{Group Name}
    \item{Units}{Unit number}
    \item{Ben}{Ben's Ratings}
    \item{Gerry}{Gerry's Ratings}
 }   
The first 2 columns "Group" and "Units" play a descriptive role here and are not used by any fucntion included in this
package. One will typically use cac.ben.gerry[,c(3,4)] or cac.ben.gerry[,c("Ben","Gerry")] as input 
dataset.}
\usage{
cac.ben.gerry
}
\description{
This dataset contains ratings that 2 raters named Ben and Gerry assigned to 12 units distributed in 2 groups 
"G1" and "G2".
}
\keyword{datasets}
