% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cac.dist4cat_doc.R
\docType{data}
\name{cac.dist4cat}
\alias{cac.dist4cat}
\title{Distribution of 4 raters by Category and Subject - Subjects allocated in 2 groups A and B.}
\format{This dataset has 15 rows (for the 15 subjects) and 4 columns. Only the last 3 columns representing the
categories into which subjects are classified are used in the calculations - unless the sub-group analysis is required.    
 \describe{
    \item{Group}{This variable repsents the subject number.}
    \item{a}{category a}
    \item{b}{Category b}
    \item{c}{Category c}
 }}
\usage{
cac.dist4cat
}
\description{
This dataset summarizes the ratings assigned by 4 raters who classified 15 subjects into one of 3 categories
named "a", "b", and "c".
}
\keyword{datasets}
