% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib.6raters_doc.R
\docType{data}
\name{distrib.6raters}
\alias{distrib.6raters}
\title{Distribution of 6 psychiatrists by Subject/patient and diagnosis Category.}
\format{
This dataset has 15 rows (for the 15 subjects) and 7 columns. Only the last 6 columns representing the
categories into which subjects are classified are used in the calculations.    
 \describe{
    \item{Subject}{This variable repsents the subject number.}
    \item{Personality.Disorder}{Personality disorder category}
    \item{Schizophrenia}{Schizophrenia Category}
    \item{Neurosis}{Neurosis category}
    \item{Other}{"Other" category}
 }
}
\source{
Fleiss, J. L. (1971). Measuring nominal scale agreement among many raters, \emph{Psychological Bulletin}, 76, 
378-382.
}
\usage{
distrib.6raters
}
\description{
This dataset summarizes the ratings assigned by 6 psychiatrists classifying 15 patients into one of five categories
named "Depression", "Personal Disorder", "Schizophrenia", "Neurosis" and "Other".
}
\keyword{datasets}
