% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.r
\name{fleiss.bf}
\alias{fleiss.bf}
\title{Computing Fleiss Benchmark Scale Membership Probabilities}
\usage{
fleiss.bf(coeff, se, BenchDF = fleiss)
}
\arguments{
\item{coeff}{A mandatory parameter representing the estimated value of an agreement coefficient.}

\item{se}{A mandatory parameter representing the agreement coefficient standard error.}

\item{BenchDF}{An optional parameter that is a 3-column data frame containing the Fleiss' benchmark scale information. 
The 3 columns are the interval lower bound, upper bound, and their interpretation. The default value is a small 
file contained in the package and named \emph{fleiss.RData}, which describes the fleiss' scale intervales and their
interpretation.}
}
\value{
A one-column matrix containing the membership probabilities (c.f. \url{http://agreestat.com/research_papers/inter-rater\%20reliability\%20study\%20design1.pdf})
}
\description{
Computing Fleiss Benchmark Scale Membership Probabilities
}
