% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fleiss_doc.R
\docType{data}
\name{fleiss}
\alias{fleiss}
\title{Dataset describing Fleiss' Benchmarking Scale}
\format{Each row of this dataset describes an interval and the interpretation of the magnitude it represents.
 \describe{
    \item{lb.FL}{The interval lower bound}
    \item{ub.FL}{The interval upper bound}
    \item{interp.FL}{The interpretation}
 }}
\source{
Fleiss, J. L. (1981). Statistical Methods for Rates and Proportions. John Wiley & Sons.
}
\usage{
fleiss
}
\description{
This dataset contains information describing Fleiss' scale for benchmarking chance-corrected agreement 
coefficients such as Gwet AC1/AC2, Kappa and many others.
}
\keyword{datasets}
