% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs.data_doc.R
\docType{data}
\name{freqs.data}
\alias{freqs.data}
\title{Distribution of 10 subjects by rater (Ben and Gerry) and by category.}
\format{
Each row of this dataset describes a unique pair of categories and the number
of subjects that both Ben and Gerry classified into the 2 categories respectively.
 \describe{
    \item{Ben}{Categories that Ben used to classify the subjects.}
    \item{Gerry}{Categories that Gerry used to classify the subjects}
    \item{n}{Number of subjects that Ben and Gerry classified into the associated pair of categories.}
 }
}
\source{
N/A.
}
\usage{
freqs.data
}
\description{
This dataset contains rating data collected from 10 human subjects by 2 raters named
Ben and Gerry. While the 5 categories a, b, c, d, and e are available for rating, Gerry
never used category a.
}
\keyword{datasets}
