% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agree.coeff3.dist.r
\name{gwet.ac1.dist}
\alias{gwet.ac1.dist}
\title{Gwet's AC1/AC2 agreement coefficient among multiple raters (2, 3, +) when the input dataset is the distribution of raters by subject and category.}
\source{
Gwet, K. L. (2008). ``Computing inter-rater reliability and its variance in the presence of high agreement," 
\emph{British Journal of Mathematical and Statistical Psychology}, 61, 29-48.
}
\usage{
gwet.ac1.dist(ratings, weights = "unweighted", categ = NULL,
  conflev = 0.95, N = Inf)
}
\arguments{
\item{ratings}{An \emph{nxq} matrix / data frame containing the distribution of raters by subject and category. Each cell \emph{(i,k)} contains the number of raters who classsified subject \emph{i} into category \emph{k}.}

\item{weights}{is an optional parameter that is either a string variable or a matrix. The string describes one of the predefined 
weights and must take one of the values ("quadratic", "ordinal", "linear", "radical", "ratio", "circular", "bipolar"). 
If this parameter is a matrix then it must be a square matri qxq where q is the number of posssible categories where a subject 
can be classified. If some of the q possible categories are not used, then it is strobgly advised to specify the complete list of 
possible categories as a vector in parametr categ. Otherwise, only the categories reported will be used.}

\item{categ}{An optional parameter representing all categories available to raters during the experiment. This parameter may be useful if 
some categories were not used by any rater inspite of being available to the raters.}

\item{conflev}{An optional parameter representing the confidence level associated with the confidence interval. Its default value is 0.95.}

\item{N}{An optional parameter representing the population size (if any). It may be use to perform the final population correction to the variance.  Its default value is infinity.}
}
\value{
A vector containing the following information: pa(the percent agreement),pe(the percent chance agreement),
coeff(Gwet's AC1 or AC2 dependending on whether weights are used or not),stderr(the standard error of Gwet's coefficient),
conf.int(the confidence interval of Gwet's coefficient), p.value(the p-value of Gwet's coefficient),coeff.name (AC1/AC2).
}
\description{
Gwet's AC1/AC2 agreement coefficient among multiple raters (2, 3, +) when the input dataset is the distribution of raters by subject and category.
}
