% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agree.coeff2.r
\name{kappa2.table}
\alias{kappa2.table}
\title{Kappa coefficient for 2 raters}
\usage{
kappa2.table(ratings, weights = identity.weights(1:ncol(ratings)),
  conflev = 0.95, N = Inf)
}
\arguments{
\item{ratings}{A square table of ratings (assume no missing ratings).}

\item{weights}{An optional matrix that contains the weights used in the weighted analysis.}

\item{conflev}{An optional confidence level for confidence intervals. The default value is the traditional 0.95.}

\item{N}{An optional population size.  The default value is infinity.}
}
\value{
A data frame containing the following 5 variables: coeff.name coeff.val coeff.se coeff.ci coeff.pval.
}
\description{
Kappa coefficient for 2 raters
}
