% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landis.koch_doc.R
\docType{data}
\name{landis.koch}
\alias{landis.koch}
\title{Dataset describing the Landis & Koch Benchmarking Scale}
\format{
Each row of this dataset describes an interval and the interpretation of the magnitude it represents.
 \describe{
    \item{lb.LK}{The interval lower bound}
    \item{ub.LK}{The interval upper bound}
    \item{interp.LK}{The interpretation}
 }
}
\source{
Landis, J.R. & Koch G. (1977). The measurement of observer agreement for categorical data, \emph{Biometrics}, 
33, 159-174.
}
\usage{
landis.koch
}
\description{
This dataset contains information describing the Landis & Koch scale for benchmarking chance-corrected agreement 
coefficients such as Gwet AC1/AC2, Kappa and many others.
}
\keyword{datasets}
