% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.gen.r
\name{linear.weights}
\alias{linear.weights}
\title{Function for computing the Linear Weights}
\usage{
linear.weights(categ)
}
\arguments{
\item{categ}{A mandatory parameter representing the vector of all possible ratings.}
}
\value{
A square matrix of quadratic weights to be used for calculating the weighted coefficients.
}
\description{
Function for computing the Linear Weights
}
