% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agree.coeff3.raw.r
\name{trim}
\alias{trim}
\title{An r function for trimming leading and trealing blanks}
\usage{
trim(x)
}
\arguments{
\item{x}{is a string variable.}
}
\value{
A string variable where leading and trealing blanks are trimmed.
}
\description{
An r function for trimming leading and trealing blanks
}
