% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x.dist10x5_doc.r
\docType{data}
\name{x.dist10x5}
\alias{x.dist10x5}
\title{Dataset of categorical ratings assigned to 10 subjects and presented in the
form of a distribution of 4 raters by subject and category}
\format{
A data frame of 10 rows and 6 columns.
integers.
 \describe{
    \item{subject}{Patient's identifier}
    \item{q1}{Number of raters who classified the subject into category q1}
    \item{q2}{Number of raters who classified the subject into category q2}
    \item{q3}{Number of raters who classified the subject into category q3}
    \item{q4}{Number of raters who classified the subject into category q4}
    \item{q5}{Number of raters who classified the subject into category q5}
 }
}
\source{
K. Gwet, PhD.
}
\usage{
x.dist10x5
}
\description{
This dataset shows how 4 raters classified 10 subjects into 5 categories
labeled as q1, q2, q3, q4 and q5. Each record is associated with a subject
and shows how the 4 raters are distributed across the 5 categories they
assigned the subject to.
}
\keyword{datasets}
