% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x.dist6x5psy_doc.r
\docType{data}
\name{x.dist6x5psy}
\alias{x.dist6x5psy}
\title{Dataset showing how 6 psychiatrists classified 15 patients by their mental
health condition.}
\format{
A data frame of 15 rows and 6 columns.
integers.
 \describe{
    \item{Subject}{Patient's identifier}
    \item{Depression}{Number of raters to have diagnosed the patient with depression}
    \item{Personality.Disorder}{Number of raters to have diagnosed the patient with personality disorder}
    \item{Schizophrenia}{Number of raters to have diagnosed the patient with schizophrenia}
    \item{Neurosis}{Number of raters to have diagnosed the patient with neurosis}
    \item{Other}{Number of raters to have diagnosed the patient with other condition}
 }
}
\source{
K. Gwet, PhD.
}
\usage{
x.dist6x5psy
}
\description{
Dataset of 15 psychiatric patients where each row identifies a patient and
shows how the 6 psychiatrists distrubte across 5 mental conditions. These
5 conditions are Depression, personality disorder, Schizophrenia, Neurosis,
and Other.
}
\keyword{datasets}
