% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x.raw10x4_doc.r
\docType{data}
\name{x.raw10x4}
\alias{x.raw10x4}
\title{Raw categorical ratings assigned to 10 subjects by 4 raters}
\format{
A data frame of 10 rows and 5 columns.
integers.
 \describe{
    \item{subject}{Patient's identifier}
    \item{rater1}{Category into which rater1 classified the subject}
    \item{rater2}{Category into which rater2 classified the subject}
    \item{rater3}{Category into which rater3 classified the subject}
    \item{rater4}{Category into which rater4 classified the subject}
 }
}
\source{
K. Gwet, PhD.
}
\usage{
x.raw10x4
}
\description{
Dataset of 10 subjects and the categorical ratings assigned to them by 4 
raters. The 5 categories available to raters are labeled as 1, 2, 3, 4, 5.
}
\keyword{datasets}
