% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agree.coeff2.r
\name{scott2.table}
\alias{scott2.table}
\title{Scott's coefficient for 2 raters}
\usage{
scott2.table(ratings, weights = identity.weights(1:ncol(ratings)),
  conflev = 0.95, N = Inf)
}
\arguments{
\item{ratings}{A square table of ratings (assume no missing ratings).}

\item{weights}{An optional matrix that contains the weights used in the weighted analysis. By default, this parameter contaings the identity weight matrix, which leads to the unweighted analysis.}

\item{conflev}{An optional parameter that specifies the confidence level used for constructing confidence intervals. By default the function assumes the standard value of 95\%.}

\item{N}{An optional parameter representing the finite population size if any. It is used to perform the finite population correction to the standard error. It's default value is infinity.}
}
\value{
A data frame containing the following 5 variables: coeff.name coeff.val coeff.se coeff.ci coeff.pval.
}
\description{
Scott's coefficient for 2 raters
}
